+++ firebase-app/firebase.json
@@ -13,8 +13,6 @@
   },
   "functions": {
     "predeploy": [
-      "npm --prefix \"$RESOURCE_DIR\" run lint",
-      "npm --prefix \"$RESOURCE_DIR\" run build"
     ]
   }
 }
