--- firebase-app/functions/README.md
@@ -1,16 +1,0 @@
-# Setup environment variables
-
-``` console
-# Setup
-$ firebase functions:config:set webhook.target_user_id="User ID"
-
-
-# Check current configuration
-$ firebase functions:config:get
-```
-
-# Deploy function
-
-``` console
-$ npm run deploy
-```
