+++ firebase-app/functions/package.json
@@ -1,10 +1,9 @@
 {
   "name": "functions",
+  "description": "Cloud Functions for Firebase",
   "scripts": {
-    "lint": "tslint --project tsconfig.json",
-    "build": "tsc",
-    "serve": "npm run build && firebase serve --only functions",
-    "shell": "npm run build && firebase functions:shell",
+    "serve": "firebase serve --only functions",
+    "shell": "firebase functions:shell",
     "start": "npm run shell",
     "deploy": "firebase deploy --only functions",
     "logs": "firebase functions:log"
@@ -12,15 +11,12 @@
   "engines": {
     "node": "10"
   },
-  "main": "lib/index.js",
   "dependencies": {
     "@google-cloud/language": "^3.6.3",
     "firebase-admin": "^8.6.0",
     "firebase-functions": "^3.3.0"
   },
   "devDependencies": {
-    "tslint": "^5.12.0",
-    "typescript": "^3.2.2",
     "firebase-functions-test": "^0.1.6"
   },
   "private": true
